+++ plugin/core/gvrender_core_mp.c
@@ -11,7 +11,8 @@
 /* FIXME - incomplete replacement for codegen */
 
 #include "config.h"
-
+#include <assert.h>
+#include <limits.h>
 #include <math.h>
 #include <stdarg.h>
 #include <stdlib.h>
@@ -21,6 +22,7 @@
 #include <io.h>
 #endif
 
+#include <cgraph/prisize_t.h>
 #include <cgraph/unreachable.h>
 #include <common/macros.h>
 #include <common/const.h>
@@ -399,8 +401,7 @@ static void mp_polygon(GVJ_t * job, pointf * A, int n, int filled)
     mpptarray(job, A, n, 1);        /* closed shape */
 }
 
-static void mp_polyline(GVJ_t * job, pointf * A, int n)
-{
+static void mp_polyline(GVJ_t *job, pointf *A, size_t n) {
     obj_state_t *obj = job->obj;
 
     int object_code = 2;        /* always 2 for polyline */
@@ -418,16 +419,17 @@ static void mp_polyline(GVJ_t * job, pointf * A, int n)
     int radius = 0;
     int forward_arrow = 0;
     int backward_arrow = 0;
-    int npoints = n;
+    const size_t npoints = n;
 
     mp_line_style(obj, &line_style, &style_val);
 
     gvprintf(job,
-            "%d %d %d %.0f %d %d %d %d %d %.1f %d %d %d %d %d %d\n",
+            "%d %d %d %.0f %d %d %d %d %d %.1f %d %d %d %d %d %" PRISIZE_T "\n",
             object_code, sub_type, line_style, thickness, pen_color,
             fill_color, depth, pen_style, area_fill, style_val, join_style,
             cap_style, radius, forward_arrow, backward_arrow, npoints);
-    mpptarray(job, A, n, 0);        /* open shape */
+    assert(n <= INT_MAX);
+    mpptarray(job, A, (int)n, 0); // open shape
 }
 
 gvrender_engine_t mp_engine = {