+++ plugin/core/gvrender_core_ps.c
@@ -382,17 +382,14 @@ static void psgen_polygon(GVJ_t * job, pointf * A, int n, int filled)
     }
 }
 
-static void psgen_polyline(GVJ_t * job, pointf * A, int n)
-{
-    int j;
-
+static void psgen_polyline(GVJ_t *job, pointf *A, size_t n) {
     if (job->obj->pencolor.u.HSVA[3] > .5) {
         ps_set_pen_style(job);
         ps_set_color(job, &(job->obj->pencolor));
 	gvputs(job, "newpath ");
 	gvprintpointf(job, A[0]);
 	gvputs(job, " moveto\n");
-        for (j = 1; j < n; j++) {
+        for (size_t j = 1; j < n; j++) {
 	    gvprintpointf(job, A[j]);
 	    gvputs(job, " lineto\n");
 	}